Reworked Component ID
Component IDs are now an enum with each component type, this makes reordering the IDs much much easier.

Mass component was removed and bundled together with the renamed force receiver, Rigidbody.

Unnecessary includes were removed as well.

+++ RaylibTest/AccelerationSystem.cpp
@@ -1,7 +1,5 @@
 #include "AccelerationSystem.h"
-#include "Scene.h"
 
-#include <raymath.h>
 /*
 ===================================================================================================
  Public Functions
@@ -9,7 +7,7 @@
 */
 void AccelerationSystem::Update(Scene& scene, std::vector<Acceleration>& accelerations, float deltaTime)
 {
-	int requiredComponentsMask = 1 << Velocity::ID;
+	int requiredComponentsMask = 1 << static_cast<int>(Velocity::ID);
 
 	for (auto& acceleration : accelerations)
 	{
