+++ RaylibTest/SpeedLimiterSystem.cpp
@@ -1,7 +1,5 @@
 #include "SpeedLimiterSystem.h"
-#include "Scene.h"
 
-#include <raymath.h>
 /*
 ===================================================================================================
  Public Functions
@@ -9,7 +7,7 @@
 */
 void SpeedLimiterSystem::Update(Scene& scene, std::vector<SpeedLimiter>& speedLimiters)
 {
-	int requiredComponentsMask = (1 << Velocity::ID) | (1 << Acceleration::ID);
+	int requiredComponentsMask = (1 << static_cast<int>(Velocity::ID)) | (1 << static_cast<int>(Acceleration::ID));
 
 	for (auto& speedLimiter : speedLimiters)
 	{
@@ -22,9 +20,6 @@ void SpeedLimiterSystem::Update(Scene& scene, std::vector<SpeedLimiter>& speedLi
 		// if at max velocity
 		if (Vector2Length(velocity.velocity) >= speedLimiter.maxVelocity)
 		{
-
-
-
 			// cancel out acceleration that is not working to decrease velocity
 			if (velocity.velocity.x * acceleration.acceleration.x > 0.0f)
 				acceleration.acceleration.x = 0.0f;
